Move contents from package [...].client.ui.ontology.search to [...].client.ui.portlet.bioportal.imports and update dependencies
