+++ src/main/java/edu/stanford/bmir/protege/web/client/ui/portlet/bioportal/imports/BioPortalSearchComponent.java
@@ -1,4 +1,4 @@
-package edu.stanford.bmir.protege.web.client.ui.ontology.search;
+package edu.stanford.bmir.protege.web.client.ui.portlet.bioportal.imports;
 
 import com.google.common.base.Optional;
 import com.google.gwt.core.client.GWT;
